+++ example/android/app/src/main/kotlin/com/authentrend/flutter_passkey_example/MainActivity.kt
@@ -2,5 +2,4 @@ package com.authentrend.flutter_passkey_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()