+++ test/flutter_passkey_test.dart
@@ -1,15 +1,68 @@
+import 'package:flutter/services.dart' show PlatformException;
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_passkey/flutter_passkey.dart';
 import 'package:flutter_passkey/flutter_passkey_platform_interface.dart';
 import 'package:flutter_passkey/flutter_passkey_method_channel.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-class MockFlutterPasskeyPlatform
+class MockFlutterPasskeyPlatformNull
     with MockPlatformInterfaceMixin
     implements FlutterPasskeyPlatform {
+  @override
+  Future<String?> getPlatformVersion() => Future.value(null);
 
   @override
-  Future<String?> getPlatformVersion() => Future.value("Android 13");
+  Future<String?> createCredential(String options) => Future.value(null);
+
+  @override
+  Future<String?> getCredential(String options) => Future.value(null);
+}
+
+class MockFlutterPasskeyPlatformAndroidOk
+    with MockPlatformInterfaceMixin
+    implements FlutterPasskeyPlatform {
+  @override
+  Future<String?> getPlatformVersion() => Future.value('Android 9');
+
+  @override
+  Future<String?> createCredential(String options) => Future.value("");
+
+  @override
+  Future<String?> getCredential(String options) => Future.value("");
+}
+
+class MockFlutterPasskeyPlatformAndroidKo
+    with MockPlatformInterfaceMixin
+    implements FlutterPasskeyPlatform {
+  @override
+  Future<String?> getPlatformVersion() => Future.value('Android 5');
+
+  @override
+  Future<String?> createCredential(String options) => Future.value("");
+
+  @override
+  Future<String?> getCredential(String options) => Future.value("");
+}
+
+class MockFlutterPasskeyPlatformIosOk
+    with MockPlatformInterfaceMixin
+    implements FlutterPasskeyPlatform {
+  @override
+  Future<String?> getPlatformVersion() => Future.value('iOS 15');
+
+  @override
+  Future<String?> createCredential(String options) => Future.value("");
+
+  @override
+  Future<String?> getCredential(String options) => Future.value("");
+}
+
+class MockFlutterPasskeyPlatformIosKo
+    with MockPlatformInterfaceMixin
+    implements FlutterPasskeyPlatform {
+  @override
+  Future<String?> getPlatformVersion() => Future.value('iOS 14');
 
   @override
   Future<String?> createCredential(String options) => Future.value("");
@@ -19,33 +72,61 @@ class MockFlutterPasskeyPlatform
 }
 
 void main() {
-  final FlutterPasskeyPlatform initialPlatform = FlutterPasskeyPlatform.instance;
+  final FlutterPasskeyPlatform initialPlatform =
+      FlutterPasskeyPlatform.instance;
 
   test('$MethodChannelFlutterPasskey is the default instance', () {
     expect(initialPlatform, isInstanceOf<MethodChannelFlutterPasskey>());
   });
 
-  test('isSupported', () async {
+  test('isSupportedNull', () async {
     FlutterPasskey flutterPasskeyPlugin = FlutterPasskey();
-    MockFlutterPasskeyPlatform fakePlatform = MockFlutterPasskeyPlatform();
-    FlutterPasskeyPlatform.instance = fakePlatform;
+    FlutterPasskeyPlatform.instance = MockFlutterPasskeyPlatformNull();
+
+    expect(await flutterPasskeyPlugin.isSupported(), false);
+  });
+
+  test('isSupportedAndroidTrue', () async {
+    FlutterPasskey flutterPasskeyPlugin = FlutterPasskey();
+    FlutterPasskeyPlatform.instance = MockFlutterPasskeyPlatformAndroidOk();
 
     expect(await flutterPasskeyPlugin.isSupported(), true);
   });
 
-  test('createCredential', () async {
+  test('isSupportedAndroidFalse', () async {
     FlutterPasskey flutterPasskeyPlugin = FlutterPasskey();
-    MockFlutterPasskeyPlatform fakePlatform = MockFlutterPasskeyPlatform();
-    FlutterPasskeyPlatform.instance = fakePlatform;
+    FlutterPasskeyPlatform.instance = MockFlutterPasskeyPlatformAndroidKo();
 
-    expect(await flutterPasskeyPlugin.createCredential(""), "");
+    expect(await flutterPasskeyPlugin.isSupported(), false);
   });
 
-  test('getCredential', () async {
+  test('isSupportedIosTrue', () async {
     FlutterPasskey flutterPasskeyPlugin = FlutterPasskey();
-    MockFlutterPasskeyPlatform fakePlatform = MockFlutterPasskeyPlatform();
-    FlutterPasskeyPlatform.instance = fakePlatform;
+    FlutterPasskeyPlatform.instance = MockFlutterPasskeyPlatformIosOk();
 
-    expect(await flutterPasskeyPlugin.getCredential(""), "");
+    expect(await flutterPasskeyPlugin.isSupported(), true);
+  });
+
+  test('isSupportedIosFalse', () async {
+    FlutterPasskey flutterPasskeyPlugin = FlutterPasskey();
+    FlutterPasskeyPlatform.instance = MockFlutterPasskeyPlatformIosKo();
+
+    expect(await flutterPasskeyPlugin.isSupported(), false);
+  });
+
+  test('createCredentialNull', () async {
+    FlutterPasskey flutterPasskeyPlugin = FlutterPasskey();
+    FlutterPasskeyPlatform.instance = MockFlutterPasskeyPlatformNull();
+
+    final future = flutterPasskeyPlugin.createCredential('');
+    await expectLater(future, throwsA(isA<PlatformException>()));
+  });
+
+  test('getCredentialNull', () async {
+    FlutterPasskey flutterPasskeyPlugin = FlutterPasskey();
+    FlutterPasskeyPlatform.instance = MockFlutterPasskeyPlatformNull();
+
+    final future = flutterPasskeyPlugin.getCredential('');
+    await expectLater(future, throwsA(isA<PlatformException>()));
   });
 }